\name{predict_SARMAX.refinement}
\alias{predict_SARMAX.refinement}
\title{Performs prediction for \code{SARMAX.refinement} function.}
\description{\code{predict_SARMAX.refinement} is a function for predictions from a \code{SARMAX.refinement} function.}

\usage{
predict_SARMAX.refinement(fit,newx)
}

\arguments{
\item{fit}{A model object from \code{SARMAX.refinement} for which prediction is desired.}
\item{newx}{Vector of new values at which predictions are performed.}
}
\author{
Zanghi Federico \email{federico.zanghi.11@gmail.com}}

\keyword{predict_SARMAX.refinement}
